TODO should be followed by task number
Summary:
There should be a rule in flint, our open source lint tool for C++, that requires comments containing the word 'TODO' to be followed by a task number. Both 'TODO' and 'Txxxxxx', where 'xxxxxx' is the task number, should be in the same line.

I'm not sure what's the proper way to handle/check comments. It seems to me
that tokenizing step munches all comments. In the current version, tokenizing step is skipped and the raw content of the file is directly checked.

Any feedback for this is highly appreciated.

Test Plan:
1. Add unit tests.
2. Run the rule on the whole code base, and eyeball the resulting
compiling errors.

Reviewed By: marcelo.juchem@fb.com

FB internal diff: D1338630

Tasks: 4064698