Implement header 'using namespace' check
It's pretty much the exact same logic as for static at global/namespace
scope.  Instead of 'static', we just need to check for 'using
namespace', so I included this check into the NamespaceScopedStatics
check.

Note that with this change the codebase no longer lints perfectly, since
we have 'using namespace std;' declarations in our headers.