Fix single-arg "explicit constexpr" constructor warning
Summary:
The presence of "constexpr" between "explicit" and the class name was
confusing flint. The C++ linter didn't have this bug since it bailed
entirely as soon as it encountered "explicit". Just skip the constexpr
keyword if we encounter it.

I added the test to the C++ codebase as well, for giggles.

Test Plan: new test

Reviewed By: njormrod@fb.com

Subscribers: louisk

FB internal diff: D1389912