Fixes default move constructor error in flint
Summary:
When the move constructor is set to default, flint
currently throws an error, stating that the move constructor should
be declared noexcept. However, this is not possible when the
constructor is set to default. Therefore, this patch checks to
see if the move constructor is the default, and then avoids
throwing the error if it is in fact the default.

Test Plan:
Added a test case to the flint tests which mimics
the error experienced in D1433822. All of the tests passed.

Reviewed By: njormrod@fb.com

Subscribers: louisk, njormrod

FB internal diff: D1433889