More specific error message from flint tokenizer in the case of illegal characters, and populates token structs with associated fileName
Summary: When flint encounters an illegal character, it will now print the file and line number location of that character rather than just the character itself.

Test Plan: Introduce an illegal character (such as any non-ascii char) into a .cpp file and run flint on it. The output should indicate the file and line number of the new character that was introduced.

Reviewed By: andrei.alexandrescu@fb.com

Subscribers: seanc

FB internal diff: D1428149

Tasks: 4026899