Relax flint sleep rule to a warning in all cases
Summary: Was talking with @njormrod and we agreed that at this point it's not all that useful to have this lint rule nagging you all the time. Having it nag for the last month raised awareness, now it's time to make sure that lint messages are more signal than noise. It's an important rule, so it's still a warning and will yell at folks for new/modified lines, but it won't apply retroactively anymore.

Test Plan: fbmake runtests

Reviewed By: njormrod@fb.com

Subscribers: louisk, njormrod, ldbrandy

FB internal diff: D1945457

Signature: t1:1945457:1427391469:9fb04e13c6ebb968aba415467747142c50a843e8